Add direction indicator (#81)
* Rename marker scripts folder

* Add direction to framework

* Add script for creating direction markers

* Fix offsets and generate HQ markers

* Add short direction in dropdown

* Make preview smaller to fit direction

* Add direction markers

* Correct spelling

+++ tools/markers/create_alphanumeric_markers.py
@@ -255,7 +255,7 @@ def main(args: argparse.Namespace):
     images = create_all_images(printable_char_sets, anchors, font)
 
     print('Cleaning export folder:', clean_images)
-    if clean_images:
+    if clean_images and args.output_dir.exists():
         shutil.rmtree(args.output_dir)
 
     print('Exporting only PNG files:', png_only)
